Rejected commit: Default
git-svn-id: http://svn.l2jfree.com/l2jfree@5150 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/handler/admincommandhandlers/AdminAnnouncements.java
@@ -92,7 +92,7 @@ public class AdminAnnouncements implements IAdminCommandHandler
 		{
 			try
 			{
-				int val = new Integer(command.substring(23)).intValue();
+				int val = new Integer(command.substring(23));
 				Announcements.getInstance().delAnnouncement(val);
 				Announcements.getInstance().listAnnouncements(activeChar);
 			}
