+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/handler/admincommandhandlers/AdminLogin.java
@@ -71,8 +71,8 @@ public class AdminLogin implements IAdminCommandHandler
 				String number = st.nextToken();
 				try
 				{
-					LoginServerThread.getInstance().setMaxPlayer(new Integer(number).intValue());
-					activeChar.sendMessage("maxPlayer set to " + new Integer(number).intValue());
+					LoginServerThread.getInstance().setMaxPlayer(new Integer(number));
+					activeChar.sendMessage("maxPlayer set to " + new Integer(number));
 					showMainPage(activeChar);
 				}
 				catch (NumberFormatException e)
